Clean up useless includes

+++ libmtpython/interpreter/arguments.cpp
@@ -10,7 +10,7 @@ using namespace mtpython::objects;
 template<typename ... Args>
 static std::string string_format(const char* format, Args ... args)
 {
-	size_t size = snprintf(nullptr, 0, format, args ...) + 1;
+	int size = snprintf(nullptr, 0, format, args ...) + 1;
 	std::unique_ptr<char[]> buf(new char[size]); 
 	snprintf(buf.get(), size, format, args ...);
 	return std::string(buf.get(), buf.get() + size - 1);
